slack-bot/pyproject.toml
[project]
name = "incharj-slack-bot"
version = "0.1.0"
description = "Slack bot for Incharj AI search"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.28.0",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.0",
    "aiohttp>=3.11.0",
]
